feat(ast/estree)!: option to return JS-only AST (#9520)
Add an `astType` option to `parseSync` and `parseAsync`.

When set to `'js'`, the returned AST does not include TypeScript-related properties. When set to `'ts'`, it does.

If not specified, the option defaults to same as source type. So when parsing JS / JSX files, the AST returned does not include TS properties by default.

The motivation for allowing the user to override that behavior, is in case someone is parsing a mix of JS and TS files, and want all the ASTs in the same format. In that case they'd set `astType: 'ts'` for all files.

This is a breaking change as it alters what properties appear in the AST on JS side.