perf(mangler): use shorter `InlineString` (#9552)
`SymbolId` is a `u32`, so there are a maximum of `u32::MAX` symbols. Therefore the longest mangled symbol name is 6 characters (`xKrTKr` for `u32::MAX`).

Use an 8-byte `InlineString` for symbol names (instead of 16 bytes).

Also, enable the tests for `oxc_mangler` crate. The tests were previously failing, but we didn't notice because they were disabled.