refactor(ast_tools): add `Output::yaml_watch_list` method (#9565)
Pure refactor. Move the logic for creating a YAML watch list into it's own function.

This was originally part of a change I was working on to make NAPI parser benchmarks only run when relevant files have changed. I've now realized that is not workable because of how Codspeed works, but I think this refactor is good anyway.