fix(napi): do not support raw transfer on WASM32 (#9566)
Prevent compilation errors on 32-bit platforms by avoiding calculating `usize: 1 << 32` unless on a 64-bit platform.

`parseSync` will throw an error if called with `experimentalRawTransfer` on an unsupported system.

Add a function `rawTransferSupported` to check if raw transfer is supported on current system or not, so user can avoid these errors.