perf(allocator): remove overflow checks from `String::from_strs_array_in` (#9650)
Optimize `String::from_strs_array_in` to remove overflow checks in common cases.

`&str` has a maximum length of `isize::MAX`, so in many cases, if we communicate this constraint to the compiler, it can remove the overflow checks, as it can see that overflowing `usize` is impossible.

I was unsure if max length of `&str` was guaranteed, but wiser minds confirm it is: https://users.rust-lang.org/t/does-str-reliably-have-length-isize-max/126777