test(transformer/using): failing test for class name (#9678)
Add a failing test for `using` transform where exported class has a reference to itself within the class body, and the external class binding is mutated.

This is something of an edge case, but it'd be ideal to handle it correctly.

Babel has the same bug: https://github.com/babel/babel/issues/17172