refactor(isolated_declarations): do not store temp values in arena (#9733)
`stmts` here is an `ArenaVec<&Statement>` (`&Statement` not `Statement`). This is a temporary collection, and doesn't end up in the AST, so we shouldn't store it in the arena.

Use a `std::vec::Vec` instead of `oxc_allocator::Vec`.

This should also remove some of the lifetime problems we have in #9656.

This change may cause a small perf hit, but in my view it's still a good change. If we want to get a speed boost, a better solution would be to have a temporary "scratch space" arena which we can allocate *all* temporary values into. This would likely give us a sizeable speed boost across many parts of Oxc (https://github.com/oxc-project/backlog/issues/121).