test(transformer/using): alter test to demonstrate semantic mismatch (#9700)
Where a `for (using x of ...)` statement has a single statement as its body, `enter_for_of_statement` wraps that statement in a `BlockStatement`. This introduces a new scope. We are missing logic to re-parent any scopes within that statement to be children of the new block scope.

This PR alters one of the tests to demonstrate this problem.