refactor(allocator): improve safety of `String::from_utf8_unchecked` (#9772)
Follow-on after #9656. Add a defence against potential double-free bug in `String::from_utf8_unchecked`.

As noted in the comment, this is probably unnecessary, but it doesn't hurt to add this defensive code while new `Vec` implementation is under development. We can remove it again later when we're satisfied we have covered all bases.