perf(linter): Return early in loop in `promise/no-nesting` (#9808)
This rule currently does unnecessary work by iterating through the all
of the args of a parent promise callback even when not necessary.

The iteration is there to check if there is a reference to any of the
given args. After the first reference is found there is no need to
continue iterating through the rest of the args.

This PR makes that loop return early as soon as the first reference to
an arg is found.