fix(ast/estree): fix TS type def for `TSThisParameter` (#9942)
Follow-on after #9913.

`TSThisParameter` is a TS-only type, so its fields aren't optional the way the TS fields of JS types are.

Fix the type def by using JS versions of added fields (`False` instead of `TsFalse` etc).

Also re-order the fields, so that `name` is first. We want the object shape to be the same (same field order) for all types which are `Identifier`s in the ESTree AST.