refactor(ast/estree): expose `INCLUDE_TS_FIELDS` constant on `Serializer` (#9943)
Remove the `serialize_ts_element` method added in #9913, and instead expose an `INCLUDE_TS_FIELDS` constant on `Serializer` trait.

Before:

```rs
if let Some(this_param) = &self.0.this_param {
    seq.serialize_ts_element(this_param);
}
```

After:

```rs
if S::INCLUDE_TS_FIELDS {
    if let Some(this_param) = &self.0.this_param {
        seq.serialize_element(this_param);
    }
}
```

This has one slight advantage. Because `INCLUDE_TS_FIELDS` is a constant, compiler will easily be able to prove that the code inside `if S::INCLUDE_TS_FIELDS { ... }` is dead code in the JS-only AST serializer, and remove it.

In this specific case, the TS-only logic is trivial, so compiler would probably see it could be optimized out anyway. But we may encounter other cases where more complex TS-only code is required, and compiler might not be able to prove that it's dead code in those cases.