test(editor): add end-to-end tests for linter output (#9979)
I'm working on laying some foundations for making our editor
integrations a little bit more robust and easier to test. This adds a
new `Tester` struct, similar to the one used in the `oxc_linter` crate
which allows us to easily run the linter on a given file and assert
against a snapshot of the diagnostic reports.

This should allow us to start adding more tests to ensure that our
editor integrations don't regress on basic functionality for the linter.

For the snapshot format, I chose to use some custom formatting since it
allows us to customize the results much better and make them more
concise and easy to read.