feat(linter): add suggested fixer to `typescript/no_unnecessary_parameter_property_assignment` and fix false positive (#9973)
Hey, I just added a suggested fix to the lint rule as continuation of my
work in #9720 and #9618.
The original typescript-eslint rule isn't capable of fixing its reported
issue, but the fix should be pretty safe.
I decided to configure it as `suggestion` and not as 100% safe `fix`.
Also found one false positive while implementing the fix suggestion,
fixed it and added a corresponding passing test.