fix(editor): re-add tester and prevent empty inverted diagnostics  (#9991)
- fixes https://github.com/oxc-project/oxc/issues/9958
- reverts + fixes https://github.com/oxc-project/oxc/pull/9982

This PR fixes an issue where the "inverted diagnostics" (i.e., the hint spans converted into diagnostic reports) cause issues in VS Code because the spans are not labeled. This happens to be the case in `no-non-null-asserted-optional-chain`: there are two diagnostic spans, but neither is labeled. This ends up creating "empty" diagnostics which cause problems. So, the fix is to simply skip creating the full diagnostic report when the message content is empty. The better fix for the future is to ensure we always label spans, as this also results in more helpful error messages.

This also re-adds the testing code which was reverted previously. I've hopefully reduced the flakiness of these tests by replacing more of the file path content to a static value.