refactor(semantic): always use `SymbolFlags::Function` for function id (#7479)
close #7402
related: #7470

The purpose of this PR is to keep the SymbolFlags of the function consistent.

Always use `SymbolFlags::Function` for the function id symbol, to avoid adding extra logic to find accurate `SymbolFlags` for the function in `transformer`. This change adds a fallback redeclaration check for case `async function foo() {}; var foo;` in `SemanticBuilder::check_redeclaraion`.

No performance difference here because redeclaration is always rare.