feat(allocator): `TakeIn` trait (#9969)
Add 2 traits and implement them for AST types:

* `Dummy` trait creates a dummy node for an AST type.
* `TakeIn` replaces a node with a dummy.

Reason for 2 separate traits is that for some types we need to implement `Dummy`, but it'd be inappropriate to implement `TakeIn`. e.g.:

* Fieldless enums which are `Copy`: They should just be copied, rather than replaced.
* `Span`: Ditto.
* `Box<T>`: You should replace the *contents* of the `Box`, not the `Box` itself.

Ideally we wouldn't be implementing `TakeIn` for so many types. Some types are very expensive to generate a dummy for, so we'd ideally guide consumers away from that by not implementing `TakeIn` for those types.

But we want to replace the `TakeIn` trait in Rolldown with our own version. Because Rolldown implements `TakeIn` on many many types, it seemed to make sense to codegen `TakeIn` for pretty much all AST types, so we can migrate their code over to using our trait without too much hassle.