refactor(semantic)!: store symbol information as the first entry in `symbol_declarations` when it is redeclared (#10062)
This PR aims to solve a problem where the flags of a symbol would merge its redeclaration symbol flags, which means we can't know what the flags of the first definition is.

This PR changes to when a symbol is redeclared, store the symbol itself as the first entry of `symbol_redeclarations`, from which we can know the `flags` of every definition.