fix(codegen): prevent arithmetic overflow calculating quote for `StringLiteral`s (#10102)
Code for calculating the optimum quote character to use for `StringLiteral`s could in theory produce arithmetic overflow (panic in debug mode, malfunction in release mode) if the string is very long and contains a very large number of quote characters e.g. `"''''''''''''"` (but much much longer).

Fix that by using `i64` instead of `i32` for the counter variables.