refactor(transformer/legacy-decorator): split the transforming decorated class into two parts and run them in `exit_class` and `exit_statement` respectively (#10153)
close: #10147

Since `class-properties` will transform the class first, which causes the property will transform into `defineProperty(xxx)` and insert in the constructor. This results in the decorators being lost. This PR splits the logic into two parts, one part used to transform decorators in `exit_class`, which runs before class-properties transformation. The other part is to transform `class A {}` into `let A = class A {}` in `exit_statement`.