fix(ast/estree): Better align program span start (#10134)
This PR adjusts the start span of the program node to the first token of
the first statement in order to align with TS-EStree serialisation.
Previously the program start span was always `0`.

The diff removes many mismatches but also adds around 35 new ones. Seem
to be mostly related to strict mode directives.

For the mismatches that are added to the snapshot, I wonder if they just
happened to not be mismatches because of the fact our Program spans used
to always be 0 and the conformance tester incorrectly expected it to
also be 0. TS-eslint seems to ignore comments when calculating the
Program start span whereas our conformance snapshot cases don't ignore
them.

The below code serialises to TS-ESTree with a program start span of `16`
(see
[playground](https://typescript-eslint.io/play/#ts=5.8.2&showAST=es&fileType=.js&esQuery=N4IgZglgNgLgpgJxALhCAvkA&code=PTAEAEBcEMCcHMCmkBcoCiBlArAKAJYBmoAFJLAK6ICUoA3gL5A&eslintrc=N4KABGBEBOCuA2BTAzpAXGYBfEWg&tsconfig=N4KABGBEDGD2C2AHAlgGwKYCcDyiAuysAdgM6QBcYoEEkJemy0eAcgK6qoDCAFutAGsylBm3TgwAXxCSgA&tokens=false)).
```
// @target: ES5
if (true) {}
```

However our conformance tester works such that we expect start span to
be `0` which doesn't line up with the playground example linked. So I
think its actually our conformance test parser thats set up wrong for
the ~35 mismatches that this PR adds to the snapshot.

---------

Co-authored-by: overlookmotel <theoverlookmotel@gmail.com>