feat(editor): add named fixes for code actions (#10203)
- related to https://github.com/oxc-project/oxc/issues/10038

This allows a `Fix` to have a message associated with it as well, similar to `RuleFix`. This is used by the editor to show code actions for a given lint error. For a start, I've added a message to the `no-debugger` rule and modified a test that shows this gets an improved code action message in VS Code.