refactor(ast_tools): split `AstKind` black list into 2 lists for structs and enums (#10250)
Replace the single black list in codegen for `AstKind` into a blacklist for structs, and a whitelist for enums.

To make what types get an `AstKind` follow a consistent pattern, we need to remove all entries from both lists. Splitting the lists like this makes it easier to see what we need to do.

Does not alter which types get an `AstKind` - just a different way of listing the "special cases".