fix(linter): `jsdoc/require-param`: skip rule if any doc has `@type` tag (#10282)
- fixes https://github.com/oxc-project/oxc/issues/10253

We had a correct skip for this rule already, but it only accounted for a
single JSDoc block. Now, if there is any `@type` tag in any JSDoc for a
given function node, we will skip the rule.

I also took this opportunity to extract a diagnostic function for this
rule and simplify the labeled span creation.