feat(minifier): compress assignment to prefix increment (#10244)
Fixes one test previously commented out in #8604.

The removed match arm for `UnaryNegation` was probably unnecessary
because `UnaryExpression(UnaryNegation, NumericLiteral(1))` seems to be
evaluated to `NumericLiteral(-1)` in a preceding minifier step and all
tests are still green. But I might be wrong as this is my first
contribution to the minifier.

---------

Co-authored-by: Boshen <boshenc@gmail.com>