perf(minifier): reduce allocations (#10301)
In these 3 places, we `pop` a statement from a `Vec` and insert it elsewhere.

Previously we used `take_in_box` to get the statement first (which involves allocating a dummy node into arena) and then `pop`-ed and discarded the dummy. Instead do the `pop` first, to avoid the unnecessary allocation.

I *assume* that compiler will see that the `unreachable!` path is genuinely unreachable because we just checked the type of the statement above, and will elide that check.