fix(linter): nested configuration directory resolution (#10157)
Fixes https://github.com/oxc-project/oxc/issues/10156

I'm not very familiar with the inner workings of this project - this was
a fix based on what I could understand after exploring the codebase for
the first time. @camchenry It would be great to get your eyes on this
since you are familiar with the nested config implementation.