refactor(ast/estree): simplify serialization for `JSXOpeningFragment` (#10316)
Follow-on after #10208.

We can remove the `JSXOpeningFragmentAttributes` serializer, and use `TsEmptyArray` instead.

This results in `attributes?: []` in the TS type def, which seems reasonable since a JSX fragment can't have attributes. Really this field shouldn't exist at all! It can probably be considered a mistake in `acorn-jsx` that it does.