perf(codegen): reduce checks printing strings (#10341)
Speed up printing strings.

I had wrongly assumed that `bytes.next().unwrap_unchecked()` would skip checks, but it seems it behaves the same as `bytes.next()` - i.e. check iterator is not empty, and only advance if it's not.

Use `assert_unchecked!` instead to inform compiler that there are definitely the required number of bytes remaining in `bytes` iterator before consuming them.

This reduces instruction count considerably, and removes a branch on every byte consumed.

https://godbolt.org/z/TWzfK1eKj

The mysteries of the compiler!