feat(language_server): use linter runtime (#10268)
closes #7118

Added tests for:
- (language_server & VSCode) nested configuration
- (language_server & VSCode) nested configuration extends
- (language_server) without nested configuration

The implementation requires at the moment a feature flag for `oxc_linter`. I added documentation in `runtime.rs` to let others know why this is needed. Maybe someone can refactor it to avoid the flag. My first idea is a Trait for file system, but you guys know what to do <3

![grafik](https://github.com/user-attachments/assets/9bb3aa67-c376-4f51-8474-1d0837d78338)