perf(parser): use `ArenaVec` to store decorators (#10437)
No performance changes show in the `CodeSpeed` because we don't have a benchmark that includes decorators. According to convention, we always store child/leaf ASTs used for building ASTs in an arena to avoid extra allocations. We have got many performances from this kind of change, so I think it is good!