fix(transformer/class-properties): `private_field_count` is incorrect when class only have `accessor` with private field (#10463)
```ts
abstract class Foo {
  accessor prop: number = 1;
  static accessor prop2: number = 1;
  accessor #prop3: number = 1;
  accessor [prop4]: number = 1;
  private accessor prop5: number = 1;
  abstract accessor prop6: number;
}
```
This case only has a `#prop3` private field with an accessor. Since we haven't supported transforming `accessor` yet, the `class_details.is_transform_required` is always `false` in this `class`, which caused the `private_field_count` to be incorrect. Moving the decremental part to the `pop` side so that we can make sure the `private_props_prop` can always be handled correctly.