refactor(ast/estree): move `FunctionFormalParameters` serializer to be with other serializers for `FormalParameters` (#10496)
Pure refactor. Revert a small change made in #10462.

`FunctionFormalParameters` serializer acts on `Function`, which is not a TS type (even though the purpose of the custom serializer is to handle a TS-specific field). Keep this serializer with the rest of the serializers related to `FormalParameters`, rather than with the serializers which act on TS-only AST types.