refactor(language_server): make linter independent of `Backend` (#10497)
This PR will refactor the language server, so the Linter is independent of the language server.
The changes are needed to support multiple workspaces in the upstream PR.
There, every linter will get its own root Uri, and will report back to the `Backend`.

The `Backend` collects the different outputs from the workers and return them to the client when needed.