fix(oxlint): current dir as arg (#9382)
closes #9023

When passing `.` as an arg it would result in a path like `{cwd}/.`,
which when being used to resolve other paths obviously caused issues.
This canonicalizes any path args before doing anything else.

Adds test to verify this based on original issue. This also caused
another snapshot to change for an invalid file extension, changing from
`LintSucceeded` to `LintNoFilesFound`, which I believe is actually
correct? If you only pass files that do not exist it would make more
sense to return an error to me so I'm viewing this as a side effect fix.