test(ast/estree): skip tests which are expected not to parse (#10608)
Part of #9705.

Skip tests which are expected not to be able to be parsed.

We may need to revert this later on - I'm not convinced all of these should fail to parse - but for now I think it's useful to have a clearer view of the remaining mismatches, as those are the ones which should be our first priority to fix.